refactor: domain과 entity 패키지 분리
