+++ src/main/java/shopping/entity/product/Product.java
@@ -1,4 +1,4 @@
-package shopping.domain.product;
+package shopping.entity.product;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
