+++ src/main/java/shopping/entity/user/Password.java
@@ -1,4 +1,4 @@
-package shopping.domain.user;
+package shopping.entity.user;
 
 import shopping.infrastructure.PasswordEncoder;
