+++ src/test/java/shopping/entity/cart/CartItemTest.java
@@ -1,10 +1,10 @@
-package shopping.domain.cart;
+package shopping.entity.cart;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import shopping.domain.DomainFixture;
-import shopping.domain.product.Product;
-import shopping.domain.user.User;
+import shopping.entity.EntityFixture;
+import shopping.entity.product.Product;
+import shopping.entity.user.User;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -14,8 +14,8 @@ class CartItemTest {
     @DisplayName("장바구니 항목 수량을 1 증가할 수 있다.")
     void increaseQuantity() {
         // given
-        User user = DomainFixture.createUser();
-        Product product = DomainFixture.createProduct();
+        User user = EntityFixture.createUser();
+        Product product = EntityFixture.createProduct();
         CartItem cartItem = new CartItem(user, product, Quantity.ONE);
 
         // when
@@ -29,8 +29,8 @@ class CartItemTest {
     @DisplayName("장바구니 항목 수량을 업데이트 할 수 있다.")
     void updateQuantity() {
         // given
-        User user = DomainFixture.createUser();
-        Product product = DomainFixture.createProduct();
+        User user = EntityFixture.createUser();
+        Product product = EntityFixture.createProduct();
         CartItem cartItem = new CartItem(user, product, Quantity.ONE);
 
         // when
