+++ src/test/java/shopping/entity/cart/CartItemsTest.java
@@ -1,10 +1,11 @@
-package shopping.domain.cart;
+package shopping.entity.cart;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import shopping.domain.DomainFixture;
-import shopping.domain.product.Product;
-import shopping.domain.user.User;
+import shopping.domain.CartItems;
+import shopping.entity.EntityFixture;
+import shopping.entity.product.Product;
+import shopping.entity.user.User;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -17,8 +18,8 @@ class CartItemsTest {
     @DisplayName("장바구니에 상품을 추가한다.")
     void addCartItem() {
         // given
-        User user = DomainFixture.createUser();
-        Product product = DomainFixture.createProduct();
+        User user = EntityFixture.createUser();
+        Product product = EntityFixture.createProduct();
         CartItem item = new CartItem(user, product, Quantity.ONE);
         CartItems items = new CartItems(new ArrayList<>());
         // when
@@ -32,8 +33,8 @@ class CartItemsTest {
     @DisplayName("장바구니에 존재하는 상품을 추가하면 수량이 늘어난다.")
     void addCartItemDuplicate() {
         // given
-        User user = DomainFixture.createUser();
-        Product product = DomainFixture.createProduct();
+        User user = EntityFixture.createUser();
+        Product product = EntityFixture.createProduct();
         CartItem item = new CartItem(user, product, Quantity.ONE);
         CartItems items = new CartItems(new ArrayList<>());
 
@@ -50,8 +51,8 @@ class CartItemsTest {
     @DisplayName("장바구니에 상품이 존재하면 true를 반환한다.")
     void contains() {
         // given
-        User user = DomainFixture.createUser();
-        Product product = DomainFixture.createProduct();
+        User user = EntityFixture.createUser();
+        Product product = EntityFixture.createProduct();
         CartItem item = new CartItem(1L, user, product, Quantity.ONE);
         CartItems items = new CartItems(new ArrayList<>(List.of(item)));
 
@@ -64,8 +65,8 @@ class CartItemsTest {
     @DisplayName("장바구니에 상품이 존재하지 않으면 false를 반환한다.")
     void notContains() {
         // given
-        User user = DomainFixture.createUser();
-        Product product = DomainFixture.createProduct();
+        User user = EntityFixture.createUser();
+        Product product = EntityFixture.createProduct();
         CartItem item = new CartItem(1L, user, product, Quantity.ONE);
         CartItems items = new CartItems(new ArrayList<>());
