+++ src/test/java/shopping/repository/ProductRepositoryTest.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-import shopping.domain.product.Product;
+import shopping.entity.product.Product;
 
 import java.util.List;
