+++ src/test/java/shopping/repository/UserRepositoryTest.java
@@ -4,9 +4,9 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-import shopping.domain.user.Email;
-import shopping.domain.user.Password;
-import shopping.domain.user.User;
+import shopping.entity.user.Email;
+import shopping.entity.user.Password;
+import shopping.entity.user.User;
 import shopping.infrastructure.SHA256PasswordEncoder;
 
 import java.util.Optional;
